conclusão do módulo Clientes

+++ database/migrations/1577367333730_telefone_cliente_schema.js
@@ -14,7 +14,7 @@ class TelefoneClienteSchema extends Schema {
         .inTable("clientes")
         .onUpdate("CASCADE")
         .onDelete("CASCADE");
-      table.integer("telefone", 11).notNullable();
+      table.bigint("telefone", 11).notNullable();
       table.timestamps();
     });
   }
